+++ eagle-stores/src/context/data/data-context.js
@@ -2,13 +2,10 @@ import axios from 'axios';
 import { createContext, useContext, useEffect, useReducer } from 'react';
 import { ACTION_TYPE } from '../../utils/actionType';
 import { initialistate, DataReducer } from '../../reducer';
-import { GetCartItems, GetWishItems } from '../../services/Services';
-import { useAuth } from '../auth/auth-context';
 
 const DataContext = createContext();
 
 const DataProvider = ({ children }) => {
-  const { token } = useAuth();
   const [state, dispatch] = useReducer(DataReducer, initialistate);
 
   useEffect(() => {
@@ -28,22 +25,6 @@ const DataProvider = ({ children }) => {
         dispatch({
           type: ACTION_TYPE.INITIALIZE_CATEGORIES,
           payload: categoryResp.data.categories,
-        });
-      }
-
-      const cartResp = GetCartItems({ encodedToken: token });
-      if (cartResp.status === 200 || cartResp.status === 201) {
-        dispatch({
-          type: ACTION_TYPE.SETCART_LIST,
-          payload: { cartlist: cartResp.data.cart },
-        });
-      }
-
-      const wishResp = GetWishItems({ encodedToken: token });
-      if (wishResp.status === 200 || wishResp.status === 201) {
-        dispatch({
-          type: ACTION_TYPE.WISHLIST,
-          payload: { wishlist: wishResp.data.wishlist },
         });
       }
     })();
